api/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
psycopg2-binary==2.9.9